+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # 0.3.5
+
+- Fixed Magisk install issue
+- Added MMAR V2 to the default repos